Conflicts:
	SKZ-ejb/src/java/dhz/skz/GlavnaFasada.java
	SKZ-ejb/src/java/dhz/skz/MainTimerBean.java
	SKZ-ejb/src/java/dhz/skz/aqdb/facades/PodatakSiroviFacadeLocal.java
	SKZ-ejb/src/java/dhz/skz/citaci/CitacMainBean.java
	SKZ-ejb/src/java/dhz/skz/citaci/mlu/MLULoggerBean.java
	SKZ-ejb/src/java/dhz/skz/citaci/weblogger/WebloggerCitacBean.java
	SKZ-ejb/src/java/dhz/skz/citaci/weblogger/validatori/SynspecValidator.java
	SKZ-ejb/src/java/dhz/skz/citaci/weblogger/validatori/ValidatorFactory.java
	SKZ-lib/src/dhz/skz/DiseminacijaGlavniBeanRemote.java
	SKZ-war/src/java/dhz/skz/beans/ZeroSpanBean.java
	SKZ-war/src/java/dhz/skz/rest/SiroviPodaci.java

+++ SKZ-lib/src/dhz/skz/aqdb/entity/DozvoljeneGranice.java
@@ -33,8 +33,8 @@ public class DozvoljeneGranice implements Serializable {
     @EmbeddedId
     protected DozvoljeneGranicePK dozvoljeneGranicePK;
     // @Max(value=?)  @Min(value=?)//if you know range of your decimal fields consider using these annotations to enforce field validation
-    private Float min;
-    private Float max;
+    private Double min;
+    private Double max;
     @JoinColumn(name = "analiticke_metode_id", referencedColumnName = "id", insertable = false, updatable = false)
     @ManyToOne(optional = false)
     private AnalitickeMetode analitickeMetode;
@@ -64,19 +64,19 @@ public class DozvoljeneGranice implements Serializable {
         this.dozvoljeneGranicePK = dozvoljeneGranicePK;
     }
 
-    public Float getMin() {
+    public Double getMin() {
         return min;
     }
 
-    public void setMin(Float min) {
+    public void setMin(Double min) {
         this.min = min;
     }
 
-    public Float getMax() {
+    public Double getMax() {
         return max;
     }
 
-    public void setMax(Float max) {
+    public void setMax(Double max) {
         this.max = max;
     }
